backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
openai==1.12.0
Pillow==10.2.0
sqlalchemy==2.0.27
python-multipart==0.0.9
pydantic==2.6.1 